feat(client): HASHI-75 검색 페이지 퍼블리싱 구조 개선
