+++ apps/client/src/pages/search/constants/searchFilters.ts
@@ -1,7 +1,6 @@
 import type {
   FoodCategoryValue,
   SearchFilterOption,
-  SearchRestaurant,
   SearchSortValue,
 } from '@/pages/search/types'
 
@@ -25,73 +24,3 @@ export const foodCategoryOptions = [
   { label: '철판/구이류', value: 'teppanGrill' },
   { label: '기타', value: 'etc' },
 ] satisfies SearchFilterOption<FoodCategoryValue>[]
-
-export const recommendedSearchKeywords = [
-  '아끼소바',
-  '라멘',
-  '돈카츠',
-  '스시',
-] as const
-
-export const searchRestaurantFixtures = [
-  {
-    id: 'akitori-musashi-1',
-    name: '아키토리 무사시 제일은 여기까지 그러니 최대길이 이 정도로까지',
-    rating: 3.8,
-    category: 'etc',
-    tag: '아끼소바',
-    businessHours: '6/19 (금) 10:00~22:00',
-    popularity: 95,
-    keywords: ['아끼소바', '야끼소바', '오코노미야키'],
-  },
-  {
-    id: 'yakisoba-kitchen-1',
-    name: '긴자 야끼소바 키친',
-    rating: 4.6,
-    category: 'teppanGrill',
-    tag: '아끼소바',
-    businessHours: '6/19 (금) 11:30~22:30',
-    popularity: 90,
-    keywords: ['아끼소바', '야끼소바', '철판'],
-  },
-  {
-    id: 'osaka-table-1',
-    name: '오사카 테이블',
-    rating: 4.2,
-    category: 'etc',
-    tag: '아끼소바',
-    businessHours: '6/19 (금) 10:30~21:30',
-    popularity: 84,
-    keywords: ['아끼소바', '오코노미야키', '타코야키'],
-  },
-  {
-    id: 'musashi-noodle-bar-1',
-    name: '무사시 누들바',
-    rating: 4.0,
-    category: 'noodle',
-    tag: '아끼소바',
-    businessHours: '6/19 (금) 12:00~23:00',
-    popularity: 79,
-    keywords: ['아끼소바', '면류', '야끼우동'],
-  },
-  {
-    id: 'ramen-ichiban-1',
-    name: '하카타 라멘 이치방',
-    rating: 4.4,
-    category: 'noodle',
-    tag: '라멘',
-    businessHours: '6/19 (금) 11:00~21:00',
-    popularity: 88,
-    keywords: ['라멘', '면류'],
-  },
-  {
-    id: 'sushi-haru-1',
-    name: '스시 하루',
-    rating: 4.7,
-    category: 'sushiSashimi',
-    tag: '스시',
-    businessHours: '6/19 (금) 12:00~22:00',
-    popularity: 91,
-    keywords: ['스시', '사시미', '초밥'],
-  },
-] satisfies SearchRestaurant[]
