+++ apps/client/src/pages/search/mocks/searchContent.mock.ts
@@ -1,0 +1,6 @@
+export const recommendedSearchKeywords = [
+  '아끼소바',
+  '라멘',
+  '돈카츠',
+  '스시',
+] as const
